requirements.txt
torch
numpy
gymnasium
stable-baselines3
tqdm
ipywidgets
matplotlib
seaborn